DAS-2060: Move full chunk to test git tracking of code chunk.
